refactor(examples-next-js-pages): simplify setup
