+++ examples/with-next-js-pages/src/services/github.ts
@@ -1,9 +1,8 @@
 import type { JSONValue } from 'zimic';
 
-import { waitForLoadedInterceptors } from '../../tests/interceptors';
 import environment from '../config/environment';
 
-const BASE_URL = environment.GITHUB_API_BASE_URL;
+const GITHUB_API_BASE_URL = environment.GITHUB_API_BASE_URL;
 const CACHE_STRATEGY = process.env.NODE_ENV === 'production' ? 'default' : 'no-store';
 
 export type GitHubRepository = JSONValue<{
@@ -15,13 +14,11 @@ export type GitHubRepository = JSONValue<{
 }>;
 
 export async function fetchGitHubRepository(ownerName: string, repositoryName: string) {
-  await waitForLoadedInterceptors();
-
   try {
     const sanitizedOwnerName = encodeURIComponent(ownerName);
     const sanitizedRepositoryName = encodeURIComponent(repositoryName);
 
-    const response = await fetch(`${BASE_URL}/repos/${sanitizedOwnerName}/${sanitizedRepositoryName}`, {
+    const response = await fetch(`${GITHUB_API_BASE_URL}/repos/${sanitizedOwnerName}/${sanitizedRepositoryName}`, {
       cache: CACHE_STRATEGY,
     });
