+++ examples/with-next-js-pages/tests/interceptors/utils.ts
@@ -1,0 +1,6 @@
+import githubInterceptor, { githubFixtures } from './github';
+
+export async function loadInterceptors() {
+  await githubInterceptor.start();
+  githubFixtures.apply();
+}
